making_thumbnail 함수 하나로 취합
