+++ petweb/account/serializers/serializer_human.py
@@ -1,10 +1,9 @@
-import os
-from PIL import Image
-
 from django.contrib.auth import get_user_model
-from django.core.files.storage import default_storage
 from rest_framework import serializers
 from rest_framework.validators import UniqueValidator
+
+from utils.making_thumb import making_thumbnail
+from .. import tasks
 
 User = get_user_model()
 
@@ -77,39 +76,6 @@ class SignupSerializer(serializers.ModelSerializer):
             image=validated_data.get('image', None),
         )
 
-        # 썸네일 생성 함수
-        def making_thumbnail(instance):
-            # 참고: (StackOverFlow: https://goo.gl/d9G7V5)
-            if instance.image.name == 'placeholder/placeholder_human.png':
-                return instance
-            new_user = User.objects.get(pk=instance.pk)
-            raw_image = default_storage.open(new_user.image.name, 'rb')
-            img = Image.open(raw_image)
-            img.thumbnail((300, 300), Image.ANTIALIAS)
-
-            image_filename, image_extension = os.path.splitext(raw_image.name)
-            image_extension = image_extension.lower()
-
-            thumb_filename = image_filename + '_thumb' + image_extension
-
-            if image_extension in ['.jpg', '.jpeg']:
-                file_type = 'JPEG',
-            elif image_extension == '.gif':
-                file_type = 'GIF'
-            elif image_extension == '.png':
-                file_type = 'PNG'
-            else:
-                return False
-
-            img.save(thumb_filename, file_type[0])
-
-            ret = thumb_filename.split('/.media/')
-
-            new_user.image = ret[1]
-            new_user.save()
-
-            return new_user
-
         thumbnail_user = making_thumbnail(created_user)
 
         return thumbnail_user
@@ -176,41 +142,6 @@ class EditSerializer(serializers.ModelSerializer):
             instance.image = new_image
         # 변경된 모든 데이터를 저장한다
         instance.save()
-
-        # 썸네일 생성 함수
-        def making_thumbnail(instance):
-            # 참고: (StackOverFlow: https://goo.gl/d9G7V5)
-            if instance.image.name == 'placeholder/placeholder_human.png':
-                return instance
-            elif '_thumb' in instance.image.name:
-                return instance
-            new_user = User.objects.get(pk=instance.pk)
-            raw_image = default_storage.open(new_user.image.name, 'rb')
-            img = Image.open(raw_image)
-            img.thumbnail((300, 300), Image.ANTIALIAS)
-
-            image_filename, image_extension = os.path.splitext(raw_image.name)
-            image_extension = image_extension.lower()
-
-            thumb_filename = image_filename + '_thumb' + image_extension
-
-            if image_extension in ['.jpg', '.jpeg']:
-                file_type = 'JPEG',
-            elif image_extension == '.gif':
-                file_type = 'GIF'
-            elif image_extension == '.png':
-                file_type = 'PNG'
-            else:
-                return False
-
-            img.save(thumb_filename, file_type[0])
-
-            ret = thumb_filename.split('/.media/')
-
-            new_user.image = ret[1]
-            new_user.save()
-
-            return new_user
 
         thumbnail_user = making_thumbnail(instance)
